requirements.txt
fritzconnection==1.4.0
PyYAML==5.3.1
phonenumberslite==8.12.15